+++ src/BenchmarkDotNet/Detectors/Cpu/Windows/WmicCpuInfoParser.cs
@@ -42,7 +42,7 @@ internal static class WmicCpuInfoParser
             }
 
             if (processor.TryGetValue(WmicCpuInfoKeyNames.MaxClockSpeed, out string frequencyValue)
-                && int.TryParse(frequencyValue, out int frequency)
+                && double.TryParse(frequencyValue, out double frequency)
                 && frequency > 0)
             {
                 nominalFrequency = nominalFrequency == 0 ? frequency : Math.Min(nominalFrequency, frequency);
